Make pytest perf infra compatible with sweep perf infra
