+++ tests/sweep_framework/sweeps/reduction/argmax/argmax_output.py
@@ -9,9 +9,8 @@ import torch
 import random
 import ttnn
 from tests.sweep_framework.framework.permutations import *
-from tests.sweep_framework.sweep_utils.utils import gen_shapes, sanitize_shape_rm
+from tests.sweep_framework.sweep_utils.utils import gen_shapes, sanitize_shape_rm, profile_ttnn_call
 from tests.tt_eager.python_api_testing.sweep_tests.generation_funcs import gen_func_with_cast_tt
-from tests.ttnn.utils_for_testing import profile_ttnn_call
 from models.utility_functions import torch_random
 from tests.sweep_framework.sweep_utils.roofline_utils import get_run_return
 from loguru import logger
