+++ tests/sweep_framework/sweeps/reduction/backward/prod_bw/prod_bw.py
@@ -10,10 +10,10 @@ import pytest
 import torch
 import random
 import ttnn
-from tests.sweep_framework.sweep_utils.utils import gen_shapes, sanitize_shape_rm
+from tests.sweep_framework.sweep_utils.utils import gen_shapes, profile_ttnn_call
 from tests.tt_eager.python_api_testing.sweep_tests.generation_funcs import gen_func_with_cast_tt
 
-from tests.ttnn.utils_for_testing import check_with_pcc, profile_ttnn_call
+from tests.ttnn.utils_for_testing import check_with_pcc
 from models.utility_functions import torch_random
 from tests.sweep_framework.sweep_utils.utils import gen_pytest_parametrize_args
 
@@ -204,4 +204,4 @@ def test_prod_bw(
     assert result, msg
     logger.info(msg)
     if e2e_perf:
-        logger.info(f"Perf. metrics: {e2e_perf}")
+        logger.info(f"E2E Perf: {e2e_perf}")
