+++ tests/sweep_framework/sweeps/reduction/mean/mean.py
@@ -11,10 +11,14 @@ import pytest
 import torch
 import random
 import ttnn
-from tests.sweep_framework.sweep_utils.utils import gen_pytest_parametrize_args, gen_shapes, sanitize_shape_rm
+from tests.sweep_framework.sweep_utils.utils import (
+    gen_pytest_parametrize_args,
+    gen_shapes,
+    sanitize_shape_rm,
+    profile_ttnn_call,
+)
 from tests.tt_eager.python_api_testing.sweep_tests.generation_funcs import gen_func_with_cast_tt
 
-from tests.ttnn.utils_for_testing import profile_ttnn_call
 from models.utility_functions import torch_random
 from tests.sweep_framework.sweep_utils.roofline_utils import get_run_return
 
@@ -167,4 +171,4 @@ def test_mean(
     assert result, msg
     logger.info(msg)
     if e2e_perf:
-        logger.info(f"Perf. metrics: {e2e_perf}")
+        logger.info(f"E2E Perf: {e2e_perf}")
