+++ tests/sweep_framework/sweeps/reduction/sum.py
@@ -114,7 +114,7 @@ def test_sum(
     assert result, msg
     logger.info(msg)
     if e2e_perf:
-        logger.info(f"Perf. metrics: {e2e_perf}")
+        logger.info(f"E2E Perf: {e2e_perf}")
 
 
 @pytest.mark.parametrize(
