+++ tests/sweep_framework/sweeps/reduction/topk/topk_output.py
@@ -12,8 +12,8 @@ from tests.sweep_framework.framework.permutations import *
 from tests.sweep_framework.sweep_utils.utils import gen_shapes, sanitize_shape
 from tests.tt_eager.python_api_testing.sweep_tests.generation_funcs import gen_func_with_cast_tt
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_topk_simmilarity
-from tests.ttnn.utils_for_testing import profile_ttnn_call
 from models.utility_functions import torch_random
+from tests.sweep_framework.sweep_utils.utils import profile_ttnn_call
 from tests.sweep_framework.sweep_utils.roofline_utils import get_run_return
 from loguru import logger
