+++ tests/sweep_framework/sweeps/reduction/var/var.py
@@ -14,10 +14,9 @@ import ttnn
 from tests.sweep_framework.sweep_utils.utils import gen_shapes, sanitize_shape_rm
 from tests.tt_eager.python_api_testing.sweep_tests.generation_funcs import gen_func_with_cast_tt
 
-from tests.ttnn.utils_for_testing import profile_ttnn_call
 from models.utility_functions import torch_random
 from tests.sweep_framework.sweep_utils.roofline_utils import get_run_return
-from tests.sweep_framework.sweep_utils.utils import gen_pytest_parametrize_args
+from tests.sweep_framework.sweep_utils.utils import gen_pytest_parametrize_args, profile_ttnn_call
 
 # Override the default timeout in seconds for hang detection.
 TIMEOUT = 30
@@ -165,4 +164,4 @@ def test_var(
     assert result, msg
     logger.info(msg)
     if e2e_perf:
-        logger.info(f"Perf. metrics: {e2e_perf}")
+        logger.info(f"E2E Perf: {e2e_perf}")
